twitPopularity/requirements.txt
django
djangorestframework
django-datatable
pip
pytz
setuptools